refactor: add register-club compliance folder in components folder
