+++ packages/web/src/features/register-club/components/compliance/RulesButton.tsx
